+++ AddOns/HandyNotes_MythicPlus/config.lua
@@ -6,8 +6,9 @@ ns.defaults = {
     profile = {
         icon_scale = 1.5,
         icon_alpha = 1.0,
-        entrances = true,
-        upcoming = false,
+    },
+    global = {
+        popup_dismissed_version = nil,
     },
 }