+++ AddOns/HandyNotes_MythicPlus/core.lua
@@ -9,7 +9,6 @@ ns.HL = HL
 
 local next = next
 local GameTooltip = GameTooltip
-local HandyNotes = HandyNotes
 
 ------------------ 
 -- Tooltip handling